Enable compiling GN by default.
This removes the call to RestoreDefaultExceptionHandler on Mac. This isn't exported by base so breaks on the component build. IT doesn't matter since this is only related to breakpad which we don't use, so I just removed the call.

BUG=288991
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/25153002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9aa0bb10f4f0c035609435efb6762f673d30476b