Fix mistake in r229993.
The point of that change was to run zip/extract/compile steps in the parent of
src/, but it just ran them in src/ with a different spelling of it.

Also do this in one more place for compile which I missed.

BUG=294387
R=ilevy@google.com
TBR=ilevy (who reviewed with his @google)

Review URL: https://codereview.chromium.org/34243002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f596f6df77c57c6fe2feeef42c283387e6cb030