Remove call to update_resources.py for WebRTC
This script is about to be removed in favor of using
download_from_google_storage.py in depot tools for
the download instead (based on .sha1 files).

When https://codereview.chromium.org/60513012/ is
landed and the builders are updated to actually execute
the hooks (https://codereview.chromium.org/59793009)
the resource files will be downloaded using that instead.
Even if executing the hooks means a bit of extra execution
time for the build, this approach is simpler and means
less maintenance.

TEST=none
BUG=webrtc:2294
NOTRY=True
TBR=navabi@chromium.org

Review URL: https://codereview.chromium.org/62273004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d7dc97f787a785965ee81d486fb23f984a020c3