Support removal of DRI platform
All ozone configurations currently build DRI support, and therefore
have a dependency on the libdrm system library. This is not needed
except when ozone_platform_dri is enabled.

This allows us to drop the link to libdrm when it is not needed.

Review URL: https://codereview.chromium.org/62953003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c64a9d2756d78931a1c1c0eb1b6e020ea2aeb93e