Add a PDF component extension
This adds a PDF component extension which is used for rendering PDFs with an
out of process PDF plugin. It adds the extension as a MIME handler for pdf file
types. The extension is currently only added if the --out-of-process-pdf switch
is passed to Chrome, AND the PDF plugin is available.

BUG=303491

Review URL: https://codereview.chromium.org/96233002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ab4ef968bf9d10bd469ddb418b3e36edff16816