Fix the official Android builder's GN build.
I don't think this needs a sysroot, and compiles for ARM, so I removed the assert and use an empty sysroot. If we need the sysroot we'll discover that when we try to make a target later.

BUG=330722
TBR=jam@chromium.org

Review URL: https://codereview.chromium.org/109023006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 53d2701d3876dcef3b75d62dd06cf0cb7854f310