Use proper clang plugin name in GN build.
This uses the proper name of the Clang plugin on Linux (previously it was hardcoded to the Mac-style name).

This also fixes the path naming so the clang plugin is correct no matter where the output directory is.

R=thakis@chromium.org

Review URL: https://codereview.chromium.org/121423002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 447ca0670bf79bdf510880be0872cec459e9500e