Drop warnings from adb shell output
This is a quick and dirty fix for a problem that occurs to me when
trying to run instrumentation tests in a simulator.
For some reason 'pm path org.chromium.content_shell_apk.tests' outputs
a warning message in my environment that confuses the python scripts
and they end up calling md5sum for the root directory which is a bit
slow... The real fix would be to be able drop the stderr of the command
but it needs a lot of plumbing and probably there is not that much chance
that something like that will happen again for someone.

BUG=334439

Review URL: https://codereview.chromium.org/138143019

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58b7c48173219583b03e8dfbc0e98574735208a0