Remove Tabpose feature on mac, and supporting infrastructure (PaintAtSize)
Tabpose has been behind a flag for 3 years, and UMA shows usage is very low.
It is the unique user of the RenderWidgetSnapshotTaker, and the only user of
RWH::PaintAtSize, which we want to remove.

XIB changes:
* Remove Tabpose menu entry "Show Tab Overview..."

BUG=251960, 223336
R=cevans@chromium.org, jamesr@chromium.org, sky@chromium.org, viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/154083008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c02ab3838f371010565d7cac859dca493160a85a