Make Linux GN build's cflags match GYP's
This is for the base shared library target on desktop Linux.

The remaining difference on Linux is that GYP's base build specifies -Wno-write-strings on Linux. This dates back to the first GYP file. I don't know why we would specify this for only base, and it seems not to trigger any warning without this flag, so I didn't add it.

This changes the pkg-config invocation from --libs to --libs-only-L and --libs-only-l. The output seems slightly different (--libs was producing an extra --export-dynamic for gmodule-2.0 that aren't present when invoking the other way, which is what the GYP build does).

R=thakis@chromium.org

Review URL: https://codereview.chromium.org/165873004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7b2a6ae1c1ecaeef835b0c1880c7d46a32cdc86