Add a new configuration 'Profile' when building for iOS
When running Instruments, it is better to use a Release build, but
with a dSYM file. So introduce a new configuration 'Profile' based
on 'Release' that also generates a dSYM file.

This uses a separate configuration because Instruments needs a
real dSYM file, and generating it with dsymutil is slow. Moreover,
it is not possible to build a dSYM file from the Release build, and
profiling a Debug build is not really representative of real usage.

BUG=None

Review URL: https://codereview.chromium.org/179933002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8125fba7a7fb8b1cf6518b6a5b4db47b0e32a806