Add support for distilling arbitrary URLs in DOM Distiller Viewer.
This adds support for requesting to view arbitrary URLs in the DOM
Distiller viewer. Previously, an entry first had to be added to the
list of articles.

Also fixes the CSP for the viewer to be in the correct format.

BUG=319881
TBR=rsleevi@chromium.org,joi@chromium.org

Review URL: https://codereview.chromium.org/151003006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6a6e6a89ec2acb8a50943900710f96873ab2e2b3