Move atrace enabling code from Chrome to WebView
According to discussions, we decided to disable atrace for Chrome
because of the inaccuracy of timing in the trace due to the overhead
of atrace.

As WebView still needs atrace, move the atrace enabling code into
AwBrowserProcess.

BUG=trace-viewer:454 (https://code.google.com/p/trace-viewer/issues/detail?id=454)

Review URL: https://codereview.chromium.org/183183011

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4c062e6ece5d5795e85a5de581a53e2ac78ed785