Template invocation fixes in GN
This adds an error check when invoking templates which caused really confusing messages if the template invocation encountered an error (because we'd continue running).

Hooks up the provider for programatically defined variables in template invocations so those can be used.

Sets the current directory in a template invocation to be that of the invoking file.

No longer define the target-related programatic variables in an import. Using these in an import will give the directory relative to the import, which is probabyl not what you want.

Fix the Windows build by adding a missing library. Add a warning not to add more to the main list (this added .lib is pretty obscure).

BUG=
R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/226223006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3aa6f818740034c3fbbc2f315ff03c89a32a9831