sync: Add UI code for per type debug counters
Adds a new tab to the about:sync page to display the latest per-type
counters.  These counters will eventually replace the StatusControllers
that contain much of the same information.

This includes some changes to the C++ SyncInternalsMessageHandler to
register to observe these counters as well as some new WebUI to display
them.

There is a known bug in this code.  The StausCounter update logic has
not been implemented yet, so the item counts will always be zero.  This
will be fixed in a future CL.

BUG=349301

Review URL: https://codereview.chromium.org/290023003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b0a15ef06ba2c5d79869033943b6825634c23190