Attempt to re-enable webrtc tests which were flaky on Android bots.
There had some CLs landed before to try fixing the flakiness of those webrtc tests on Android bot. But since the flakiness only happens on bots but not locally, so I enabled it again to see if those CLs help or not.

If any of the WebRTC tests still flake on bots, this CL should be reverted.

BUG=343604
TEST=bots

Review URL: https://codereview.chromium.org/408463002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7e052bf244a4db606fb7ac51b3ba2a640170a88