Replace the use of CLD_DATA_FROM* and CLD-related ifdefs with runtime checks.
This, along with other changes described in crbug/367239, allows refactoring
of the build logic to allow per-target customization of the CLD data source.
This, in turn, allows chrome_shell to be built with statically-linked CLD even
if other things (such as the full browser) are not.

BUG=367239
TBR=toyoshim@chromium.org

Review URL: https://codereview.chromium.org/424123003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3cb7c2b55cfa24ae40cd49a1d8fd15275b55dd0b