Allow selection between libstdc++ and libc++ when targeting iOS
The recommended C++ library on iOS is libc++ but the compiler still default
to libstdc++ so add a new gyp variable "use_system_libcxx" to choose which
C++ library to use.

When building Chrome for iOS, so tool are still build on the host which
targets Mac OS X SDK 10.6 which does not supports libc++ so only enabled the
usage of libc++ when not building for the host.

The version of clang shipped on Google storage does not supports using the
system libc++ so instead use the version of clang shipped with Xcode when
using libc++.

BUG=366744

Review URL: https://codereview.chromium.org/425313002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 480c6676fb4dea61d6c896dff3ddca8fe4b869ed